Separated cluster to hub & client
